Dev mobile v1

+++ cloud_final/cloud_athentification/crypto_rojo/src/main/java/com/crypto/service/SessionUtilisateurService.java
@@ -2,14 +2,18 @@ package com.crypto.service;
 
 import com.crypto.model.SessionUtilisateur;
 import com.crypto.repository.SessionUtilisateurRepository;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
 public class SessionUtilisateurService {
+    @Autowired
     SessionUtilisateurRepository sessionUtilisateurRepository;
     public void deleteToken(String token)
     {
         SessionUtilisateur sessionUtilisateur = sessionUtilisateurRepository.getByToken(token);
-        sessionUtilisateurRepository.delete(sessionUtilisateur);
+        if (sessionUtilisateur != null){
+            sessionUtilisateurRepository.delete(sessionUtilisateur);
+        }
     }
 }
